refactor ui fixtures package structure
