+++ ui/src/integrationTest/java/de/cotto/lndmanagej/ui/BaseControllerIT.java
@@ -3,7 +3,7 @@ package de.cotto.lndmanagej.ui;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.mock.mockito.MockBean;
 
-import static de.cotto.lndmanagej.controller.dto.StatusModelFixture.STATUS_MODEL;
+import static de.cotto.lndmanagej.ui.dto.StatusModelFixture.STATUS_MODEL;
 import static org.mockito.Mockito.when;
 
 public class BaseControllerIT {
