+++ ui/src/test/java/de/cotto/lndmanagej/ui/StatusServiceImplTest.java
@@ -7,7 +7,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import static de.cotto.lndmanagej.controller.dto.StatusModelFixture.STATUS_MODEL;
+import static de.cotto.lndmanagej.ui.dto.StatusModelFixture.STATUS_MODEL;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
