+++ ui/src/testFixtures/java/de/cotto/lndmanagej/ui/dto/OpenChannelDtoFixture.java
@@ -1,6 +1,7 @@
-package de.cotto.lndmanagej.controller.dto;
+package de.cotto.lndmanagej.ui.dto;
 
-import de.cotto.lndmanagej.ui.dto.OpenChannelDto;
+import de.cotto.lndmanagej.controller.dto.BalanceInformationDto;
+import de.cotto.lndmanagej.controller.dto.PoliciesDto;
 
 import static de.cotto.lndmanagej.model.BalanceInformationFixtures.BALANCE_INFORMATION;
 import static de.cotto.lndmanagej.model.BalanceInformationFixtures.BALANCE_INFORMATION_2;
