+++ ui/src/testFixtures/java/de/cotto/lndmanagej/ui/dto/StatusModelFixture.java
@@ -1,6 +1,4 @@
-package de.cotto.lndmanagej.controller.dto;
-
-import de.cotto.lndmanagej.ui.dto.StatusModel;
+package de.cotto.lndmanagej.ui.dto;
 
 public class StatusModelFixture {
